+++ CHANGELOG.md
@@ -29,6 +29,10 @@
   ([#6764](https://github.com/mitmproxy/mitmproxy/pull/6764), @changsin)
 * Add primitive websocket interception and modification
   ([#6766](https://github.com/mitmproxy/mitmproxy/pull/6766), @errorxyz)
+* Add support for exporting websocket messages when using "raw" export.
+  ([#6767](https://github.com/mitmproxy/mitmproxy/pull/6767), @txrp0x9)
+* The "save body" feature now also includes WebSocket messages.
+  ([#6767](https://github.com/mitmproxy/mitmproxy/pull/6767), @txrp0x9)
 
 ## 07 March 2024: mitmproxy 10.2.4